chore: remove moved page files
