+++ src/pages/Live/components/SensorDashboard/index.jsx
@@ -3,7 +3,7 @@ import React, {useEffect, useMemo, useState} from "react";
 import Header from "../../../../components/Header";
 import {useLiveDevices} from "../../../../components/useLiveDevices";
 import { useLiveNow } from "../../../../hooks/useLiveNow";
-import styles from "./SensorDashboard.module.css";
+import styles from "../../../common/SensorDashboard.module.css";
 import Live from "../Live";
 import {SENSOR_TOPIC, topics} from "../../../../components/dashboard.constants";
 import {useFilters, ALL} from "../../../../context/FiltersContext";
