+++ src/pages/common/SensorDashboard.module.css
